+++ client/src/lib.rs
@@ -1477,6 +1477,14 @@ impl Client {
 
     pub fn unmount(&mut self) { self.send_msg(ClientGeneral::ControlEvent(ControlEvent::Unmount)); }
 
+    pub fn set_pet_stay(&mut self, entity: EcsEntity, stay: bool) {
+        if let Some(uid) = self.state.read_component_copied(entity) {
+            self.send_msg(ClientGeneral::ControlEvent(ControlEvent::SetPetStay(
+                uid, stay,
+            )));
+        }
+    }
+
     pub fn respawn(&mut self) {
         if self
             .state