+++ common/src/event.rs
@@ -200,6 +200,7 @@ pub enum ServerEvent {
     Mount(EcsEntity, EcsEntity),
     MountVolume(EcsEntity, VolumePos),
     Unmount(EcsEntity),
+    SetPetStay(EcsEntity, EcsEntity, bool),
     Possess(Uid, Uid),
     /// Inserts default components for a character when loading into the game
     InitCharacterData {