+++ common/systems/src/controller.rs
@@ -63,6 +63,11 @@ impl<'a> System<'a> for Sys {
                             }
                         }
                     },
+                    ControlEvent::SetPetStay(pet_uid, stay) => {
+                        if let Some(pet_entity) = read_data.id_maps.uid_entity(pet_uid) {
+                            server_emitter.emit(ServerEvent::SetPetStay(entity, pet_entity, stay));
+                        }
+                    },
                     ControlEvent::RemoveBuff(buff_id) => {
                         server_emitter.emit(ServerEvent::Buff {
                             entity,