+++ server/src/sys/agent.rs
@@ -68,9 +68,11 @@ impl<'a> System<'a> for Sys {
             read_data.light_emitter.maybe(),
             read_data.groups.maybe(),
             read_data.rtsim_entities.maybe(),
-            !&read_data.is_mounts,
-            read_data.is_riders.maybe(),
-            read_data.is_volume_riders.maybe(),
+            (
+                !&read_data.is_mounts,
+                read_data.is_riders.maybe(),
+                read_data.is_volume_riders.maybe(),
+            ),
         )
             .par_join()
             .for_each_init(
@@ -93,9 +95,7 @@ impl<'a> System<'a> for Sys {
                     light_emitter,
                     group,
                     rtsim_entity,
-                    _,
-                    is_rider,
-                    is_volume_rider,
+                    (_, is_rider, is_volume_rider),
                 )| {
                     let mut event_emitter = event_bus.emitter();
                     let mut rng = thread_rng();