+++ voxygen/src/session/mod.rs
@@ -13,14 +13,14 @@ use vek::*;
 
 use client::{self, Client};
 use common::{
-    comp,
     comp::{
+        self,
         dialogue::Subject,
         inventory::slot::{EquipSlot, Slot},
         invite::InviteKind,
         item::{tool::ToolKind, ItemDesc},
-        ChatType, Content, InputKind, InventoryUpdateEvent, Pos, PresenceKind, Stats,
-        UtteranceKind, Vel,
+        CharacterActivity, ChatType, Content, InputKind, InventoryUpdateEvent, Pos, PresenceKind,
+        Stats, UtteranceKind, Vel,
     },
     consts::MAX_MOUNT_RANGE,
     event::UpdateCharacterMetadata,
@@ -32,6 +32,7 @@ use common::{
     trade::TradeResult,
     util::{Dir, Plane},
     vol::ReadVol,
+    CachedSpatialGrid,
 };
 use common_base::{prof_span, span};
 use common_net::{msg::server::InviteAnswer, sync::WorldSyncExt};
@@ -936,6 +937,45 @@ impl PlayState for SessionState {
                                             client.mount(mountee_entity);
                                         }
                                     }
+                                }
+                            },
+                            GameInput::StayFollow if state => {
+                                let mut client = self.client.borrow_mut();
+                                let player_pos = client
+                                    .state()
+                                    .read_storage::<Pos>()
+                                    .get(client.entity())
+                                    .copied();
+
+                                let mut close_pet = None;
+                                if let Some(player_pos) = player_pos {
+                                    let positions = client.state().read_storage::<Pos>();
+                                    close_pet = client.state().ecs().read_resource::<CachedSpatialGrid>().0
+                                        .in_circle_aabr(player_pos.0.xy(), MAX_MOUNT_RANGE)
+                                        .filter(|e|
+                                            *e != client.entity()
+                                        )
+                                        .filter(|e|
+                                            matches!(client.state().ecs().read_storage::<comp::Alignment>().get(*e),
+                                                Some(comp::Alignment::Owned(owner)) if Some(*owner) == client.uid())
+                                        )
+                                        .filter(|e|
+                                            client.state().ecs().read_storage::<Is<Mount>>().get(*e).is_none()
+                                        )
+                                        .min_by_key(|e| {
+                                            OrderedFloat(positions
+                                                .get(*e)
+                                                .map_or(MAX_MOUNT_RANGE * MAX_MOUNT_RANGE, |x| {
+                                                    player_pos.0.distance_squared(x.0)
+                                                }
+                                            ))
+                                        });
+                                }
+                                if let Some(pet_entity) = close_pet && client.state().read_storage::<Is<Mount>>().get(pet_entity).is_none() {
+                                    let is_staying = client.state()
+                                        .read_component_copied::<CharacterActivity>(pet_entity)
+                                        .map_or(false, |activity| activity.is_pet_staying);
+                                    client.set_pet_stay(pet_entity, !is_staying);
                                 }
                             },
                             GameInput::Interact => {